requirements.txt
absl-py
grpcio-tools
redis